Cleanup test code
