--- scripts/ccf/az-cleanroom-aci/orchestrator/requirements.txt
@@ -1,1 +1,0 @@
-pytest
